[NO-JIRA][nx] Migrate Storybook config into backpack-storybook-host lib
Moves the root .storybook/ directory into libs/backpack-storybook-host/
as a standalone Nx project (scope:internal, type:storybook). Wires the
Storybook plugin to host stories from packages/backpack-web/src/, points
preview/themeableAttributes at the @skyscanner/backpack-web package
import, and replaces root storybook/percy-test scripts with nx targets.
Adds typecheck (tsc --build for lib + storybook configs), lint, and
visual-test targets, and registers the new project reference in
tsconfig.solutions.json.
