+++ libs/backpack-storybook-host/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "backpack-storybook-host",
+  "version": "0.0.0",
+  "private": true,
+  "description": "Internal Storybook 10 host configuration for the Backpack design system.",
+  "license": "Apache-2.0",
+  "main": "src/index.ts",
+  "exports": {
+    ".": {
+      "types": "./src/index.ts",
+      "default": "./src/index.ts"
+    }
+  },
+  "dependencies": {
+    "@skyscanner/backpack-web": "*",
+    "bpk-storybook-utils": "*"
+  }
+}
